requirements.txt
confluent_kafka
pandas
joblib
sqlalchemy
python-decouple
psycopg2-binary
